backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.0
httpx>=0.27.0
langgraph>=0.2.0

# Optional (only needed if you want to call OpenAI-style APIs via the official SDK)
openai>=1.40.0

# Optional: For web search tool (recommended for better search results)
# Install with: pip install duckduckgo-search
# duckduckgo-search>=3.9.0
